# session_4_swagger/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserLoginResponse": {
				"properties": {
					"session_token": {
						"type": "string"
					},
					"is_success": {
						"type": "boolean"
					}
				},
				"required": [
					"session_token",
					"is_success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserLoginRequest": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserRegisterResponse": {
				"properties": {
					"is_success": {
						"type": "boolean"
					}
				},
				"required": [
					"is_success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserRegisterRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserConfirmationResponse": {
				"properties": {
					"is_success": {
						"type": "boolean"
					}
				},
				"required": [
					"is_success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TodosResponse": {
				"properties": {
					"is_success": {
						"type": "boolean"
					}
				},
				"required": [
					"is_success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TodosRequest": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"session_token": {
						"type": "string"
					},
					"todo": {
						"type": "string"
					}
				},
				"required": [
					"user_id",
					"session_token",
					"todo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrmTodo": {
				"properties": {
					"todo_id": {
						"type": "number",
						"format": "double"
					},
					"user_id": {
						"type": "number",
						"format": "double"
					},
					"todo": {
						"type": "string"
					},
					"is_deleted": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"todo_id",
					"user_id",
					"todo",
					"is_deleted",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "session_4_backend_leftovers",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/dummy/DummyFunction": {
			"post": {
				"operationId": "DummyFunction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "input_name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserLoginResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserLoginRequest"
							}
						}
					}
				}
			}
		},
		"/users/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserRegisterResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserRegisterRequest"
							}
						}
					}
				}
			}
		},
		"/users/confirmation/{uuid}": {
			"get": {
				"operationId": "Confirmation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserConfirmationResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/todos/add": {
			"post": {
				"operationId": "Add",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TodosResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TodosRequest"
							}
						}
					}
				}
			}
		},
		"/todos/list": {
			"post": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/OrmTodo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TodosRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}